agent_runtime_v2/requirements.txt
semantic-kernel>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0
